4.21 (auth for blog DELETE)

+++ osa4/blog/controllers/blogs.js
@@ -34,8 +34,25 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.deleteOne({ _id: request.params.id })
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  console.log(`user ${user.id} is attempting to delete blog ${request.params.id} owned by ${blog.user._id}`)
+  console.log(`match: ${user.id == blog.user._id}`)
+  if (user.id.toString() === blog.user._id.toString()) {
+    await Blog.deleteOne({ _id: request.params.id })
+    console.log(`deleted blog ${request.params.id}`)
+    response.status(204).end()
+  } else {
+    console.log('unauthorized, did not delete')
+    return response.status(401).json({ error: 'unauthorized' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
